Move articleDetials state and fetch to ArticleDetailsComponent and pass slug prop
